fix(prix): la marge est enfin GARDEE mecaniquement, et la colonne cesse de retrecir le produit
Quatrieme revue a froid : 4 rouges, 6 jaunes. Tous fermes sauf l essai reel, qui est declare.

🔴 LE PLUS IMPORTANT : AUCUN TEST NE VOYAIT LE CABLAGE. Mesure faite avant de corriger : en remettant
`return brut;` a la place de `pricingFacture(...)`, c est-a-dire EXACTEMENT le defaut qu on venait de
corriger, les 5683 tests restaient VERTS. Les cas existants appellent les fonctions en direct : ils
prouvent qu elles calculent juste, jamais qu on les appelle. C est le piege n°4 de la skill de revue, mot
pour mot. `tests/prix-cablage.test.ts` garde desormais les trois branchements, et les trois mutations ont
ete verifiees. Le defaut a ete oublie TROIS fois de suite ; il ne peut plus l etre en silence.

🔴 ET LE NOM ETAIT LA CAUSE. `tarifsMeta` rendait des prix de VENTE depuis deux commits, et l inventaire
des consommateurs avait ete fait sur « qui appelle tarifsMeta » plutot que sur « qui affiche un prix a un
client ». Renommee `prixFactures`.

🔴 LA COLONNE DE LA MARGE RETRECISSAIT LE PRODUIT. `smallint` interdisait une marge de 120,5 %, soit
+20,5 %, banale commercialement, et l ecran la refusait sans pouvoir expliquer pourquoi puisqu elle est
DANS les bornes annoncees. La migration n etant pas appliquee, c est la COLONNE qui a ete elargie en
`numeric(6,2)` : un type choisi sans y penser ne doit pas decider de ce qu un client a le droit de
facturer. Un test relit le fichier SQL pour que la validation et la colonne ne divergent plus.

🔴 LA PHRASE QUI RECONCILIE DEUX CHIFFRES DE LA MEME PAGE DESIGNAIT LA MAUVAISE CAUSE. Elle attribuait tout
l ecart entre le cout estime et la facture Meta au tarif moyen par categorie. Avec une marge de 150, l
ecart est de 50 % et sa cause dominante n etait nommee nulle part : le client lisait deux nombres et
l ecran lui montrait un arrondi. La marge voyage desormais avec ce qu elle explique, et n est nommee que
quand elle existe.

🔴 features.md CONTREDISAIT LES ECRANS en cinq endroits, tous ecrits par mes propres commits.

Jaunes fermes : `pricingFacture` ne marge plus `service` ni `authentication` (leur prix se saisit, il ne se
derive d aucun tarif Meta, et un futur lecteur aurait eu un prix faux sans erreur), l essai reel du plan
pose maintenant une marge (sans quoi il ne verrait pas un septieme consommateur oublie, le defaut etant
INVISIBLE a marge 100), `coutParClic` rejoint le backlog avec le meme denominateur non borne, et deux
commentaires nommaient le mauvais fichier de test.

+++ db/migrations/0154_grille_prix_espace.sql
@@ -33,7 +33,13 @@
 -- poste de travail pointe sur la PRODUCTION, et une migration s'applique sur le VPS, image construite
 -- d'abord (`compose build mba-api`, puis `compose run --rm --no-deps mba-api npm run migrate`).
 
-alter table tenant_settings add column if not exists prix_marge_template     smallint     not null default 100;
+-- 🔴 `numeric(6,2)` ET PAS `smallint`, ET LE CHOIX A ETE CORRIGE AVANT L APPLICATION. Un `smallint`
+-- interdisait une marge de 120,5 %, c est-a-dire +20,5 %, qui est une marge commerciale banale. Pire : la
+-- validation applicative le refusait avec « valeur refusee » sans pouvoir expliquer pourquoi, puisque la
+-- valeur EST dans les bornes annoncees (1 a 1000). On retrecissait le produit pour tenir un type choisi
+-- sans y penser. La migration n'etant pas encore appliquee, l'elargir ne coute rien. Releve a la quatrieme
+-- revue du 2026-09-18.
+alter table tenant_settings add column if not exists prix_marge_template     numeric(6,2) not null default 100;
 alter table tenant_settings add column if not exists prix_service_centimes   numeric(6,2) not null default 2.48;
 alter table tenant_settings add column if not exists prix_service_franchise  integer      not null default 1000;
 alter table tenant_settings add column if not exists prix_service_depuis     date         not null default date '2026-10-01';
